Add files via upload
Coding Exercise - Enum Direction
In this problem, you are given an enumeration called Direction which represents the four cardinal directions: NORTH, SOUTH, EAST, WEST.

The goal is to implement the left() and right() methods that correctly return the Direction instance to the left and right, respectively, of a given Direction instance.

The directions are arranged clockwise.

Therefore, the direction to the "left" of NORTH is WEST, of WEST is SOUTH, of SOUTH is EAST, and of EAST is NORTH.

Similarly, the direction to the "right" of NORTH is EAST, of EAST is SOUTH, of SOUTH is WEST, and of WEST is NORTH.



Notes

The switch statement is used to select one of many code blocks to be executed. It's a multi-way branch statement that compares the value of a variable or expression with multiple cases.

The this keyword in Java is used inside a method or constructor to refer itself to the current object.